+++ main-service/src/main/java/ru/practicum/ewm/mainservice/user/service/UserService.java
@@ -6,7 +6,7 @@ import ru.practicum.ewm.mainservice.user.dto.UserDto;
 import java.util.List;
 
 public interface UserService {
-    List<UserDto> getUsers(List<Integer> ids, Integer from, Integer size);
+    List<UserDto> getUsers(List<Long> ids, Integer from, Integer size);
 
     UserDto registerUser(NewUserRequest newUserRequest);
